Refactor navbar and authentication components: Update image formats to WebP, adjust navbar logo size, modify vertical divider colors, and change layout heights for better responsiveness.
