+++ src/App.jsx
@@ -1,38 +1,70 @@
-import React from 'react';
-import { BrowserRouter as Router, Routes, Route } from 'react-router-dom';
-import Navbar from './components/Layout/navbar/navbar';
-import Footer from './components/Layout/footer/footer';
+import { React, useEffect, lazy, Suspense } from "react";
+import {
+  BrowserRouter as Router,
+  Routes,
+  Route,
+  useLocation,
+} from "react-router-dom";
+import Navbar from "./components/Layout/navbar/navbar";
+import Footer from "./components/Layout/footer/footer";
+import Home from "./features/Home/home";
+import SignUp from "./features/Auth/signUp/signUp";
 
-import SignIn from './features/Auth/signIn/signIn'; // ← ruta real
-import SignUp from './features/Auth/signUp/signUp'; // ← ruta real
-import Home from './features/Home/home';            // ← ruta real
-import About from './features/aboutUs/aboutUs'; // ← nueva línea
-import VerifyEmail from './features/Auth/verifyEmail/verifyEmail'; // ← nueva línea
-import VerifyCode from './features/Auth/verifyCode/verifyCode'; // ← nueva línea
-import Forgot from './features/Auth/forgot/forgot'; // ← nueva línea
+// Lazy loading para los componentes necesarios
+const SignIn = lazy(() => import("./features/Auth/signIn/signIn"));
+const About = lazy(() => import("./features/aboutUs/aboutUs"));
+const VerifyEmail = lazy(() =>
+  import("./features/Auth/verifyEmail/verifyEmail")
+);
+const VerifyCode = lazy(() => import("./features/Auth/verifyCode/verifyCode"));
+const Forgot = lazy(() => import("./features/Auth/forgot/forgot"));
+const FullScreenLoader = () => (
+  <div
+    style={{
+      height: "100vh",
+      display: "flex",
+      justifyContent: "center",
+      alignItems: "center",
+    }}
+  >
+    <div className="spinner"></div>
+  </div>
+);
 
-import './App.css';
+
+import "./App.css";
 
 function App() {
-  return (
-      <div className="app-container">
-        <Navbar />
+  const location = useLocation();
 
-        <main className="container">
+  useEffect(() => {
+    if (location.pathname === "/" || location.pathname === "/about") {
+      document.body.classList.add("home-background");
+    } else {
+      document.body.classList.remove("home-background");
+    }
+  }, [location.pathname]);
+
+  return (
+    <div className="app-container">
+      <Navbar />
+
+      <main className="container">
+        <Suspense fallback={<FullScreenLoader />}>
           <Routes>
             <Route path="/" element={<Home />} />
             <Route path="/signin" element={<SignIn />} />
             <Route path="/signup" element={<SignUp />} />
-            <Route path="/about" element={<About />} /> 
+            <Route path="/about" element={<About />} />
             <Route path="/verify-email" element={<VerifyEmail />} />
             <Route path="/verify-code" element={<VerifyCode />} />
             <Route path="/forgot-password" element={<Forgot />} />
-            {/* Puedes agregar más rutas aquí */}
           </Routes>
-        </main>
+        </Suspense>
+      </main>
 
-        <Footer />
-      </div>
+      <Footer />
+    </div>
   );
 }
