+++ src/features/Auth/forgot/forgot.css
@@ -32,7 +32,7 @@
   
   .vertical-divider {
     width: 2px;
-    background-color: black;
+    background-color: #2975bc;
     height: 100%;
   }
