+++ src/features/Auth/signIn/signIn.css
@@ -31,7 +31,7 @@
 
 .vertical-divider {
   width: 2px;
-  background-color: black;
+  background-color: #2975bc;
   height: 100%;
 }
