+++ src/features/Auth/verifyCode/verifyCode.css
@@ -20,7 +20,7 @@
 
   .vertical-divider {
     width: 2px;
-    background-color: black;
+    background-color: #2975bc;
     height: 100%;
 }
