+++ src/features/Auth/verifyEmail/verifyEmail.css
@@ -24,7 +24,7 @@
 
 .vertical-divider {
   width: 2px;
-  background-color: black;
+  background-color: #2975bc;
   height: 100%;
 }
