+++ src/features/aboutUs/aboutUs.css
@@ -1,5 +1,5 @@
 .about-container {
-    min-height: 100vh;
+    min-height: 82.4vh;
     display: flex;
     justify-content: center;
     align-items: center;
